# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 百舸API网关
  description: |
    统一百舸各模块API规范和数据结构的中间层网关服务。

    所有接口采用RPC风格，通过 `?action=XXX` 参数区分操作。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/v1
    description: 本地开发服务器
...
paths:
...
  /aihc/resource-pools/DescribeResourcePools:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeResourcePools
      operationId: describeResourcePools
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePoolListResponse"
  /aihc/resource-pools/DescribeResourcePool:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeResourcePool
      operationId: describeResourcePool
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePoolDetail"
  /aihc/resource-pools/DescribeQueues:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeQueues
      operationId: describeQueues
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueListResponse"
  /aihc/resource-pools/DescribeQueue:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeQueue
      operationId: describeQueue
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueDetail"
  /aihc/resource-pools/DescribeNodes:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeNodes
      operationId: describeNodes
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeListResponse"
  /aihc/resource-pools/DescribeNode:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeNode
      operationId: describeNode
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
  /aihc/resource-pools/DescribeResourcePoolUsage:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeResourcePoolUsage
      operationId: describeResourcePoolUsage
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePoolUsage"
...
components:
  parameters:
    RegionParam:
      name: region
      in: header
      required: true
      description: 区域代码
      schema:
        type: string
        enum:
          - bj
          - gz
          - su
          - bd
          - fwh
          - yq
        default: bd
    AkParam:
      name: ak
      in: header
      required: true
      description: 百度云 Access Key
      schema:
        type: string
    SkParam:
      name: sk
      in: header
      required: true
      description: 百度云 Secret Key
      schema:
        type: string
  schemas:
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
...
    ResourcePoolListResponse:
      type: object
      properties:
        resourcePools:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    ResourcePoolDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - selfManaged
            - managed
        status:
          type: string
        nodeCount:
          type: integer
    QueueListResponse:
      type: object
      properties:
        queues:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    QueueDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        resourcePoolId:
          type: string
    NodeListResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    NodeDetail:
      type: object
      properties:
        name:
          type: string
        instanceId:
          type: string
        status:
          type: string
        resources:
          type: object
    ResourcePoolUsage:
      type: object
      properties:
        usage:
          type: object
          properties:
            total:
              type: object
            used:
              type: object
            available:
              type: object
...